pci: multiple "number" specification for one device support
If the config.vmm.driver.pci has multiple "number" selection for one
device, previously its order had to be descending order, like this:

  device=uhci, number=1, driver=uhci, and,
  device=uhci, number=0, driver=conceal

The reason is that the number was counted during interpretation but
the interpretation was terminated when device specification was
matched.

This changeset fixes the bug.  Fixing the bug needs rewrite of
interpreting routine.  Now the pci_match_find_driver_sub() function is
simpler than before.