pci: save PCI/PCIe bridge info and use initial bus number for pci_match
A PCI/PCIe bridge has the secondary bus number that is a bus number
used by devices which connected to the PCI/PCIe bridge.  Previously
the first %02x of slot=%02x:%02x.%u in the pci_match routines meant a
bus number of devices.  However the bus number could be changed by the
guest operating system because the secondary bus number of PCI/PCIe
bridges can be changed.

Now the pci_match routines use the initial bus number that cannot be
changed by the guest operating system.  If a PCI/PCIe bridge did not
exist on start-up, the initial bus number will be -1 which is ffffffff
in %02x.