pro1000: add virtio-net emulation
A new driver option "virtio=1" enables the virtio-net emulation.  If
it is enabled, the network interface is initialized by the VMM and can
be used before starting the guest operating system, and, the guest
operating system will detect the network interface as a virtio-net
device.

Thanks to Yushi Omote <omote@osss.cs.tsukuba.ac.jp> for idea of using
virtio-net.